requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
dnspython==2.7.0
email_validator==2.2.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pydantic==2.10.1
pydantic_core==2.27.1
pymongo==4.10.1
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
